requirements.txt
amqp==5.0.1
asgiref==3.2.10
billiard==3.6.3.0
celery==5.0.0
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
Django==3.0.8
django-tables2==2.3.1
djangorestframework==3.12.1
importlib-metadata==2.0.0
kombu==5.0.2
Pillow==7.2.0
prompt-toolkit==3.0.8
pytz==2020.1
redis==3.5.3
six==1.15.0
sqlparse==0.3.1
vine==5.0.0
wcwidth==0.2.5
zipp==3.3.0
